\name{Constraints-class}
\alias{Constraints}
\alias{Constraints-class}
\alias{Unconstrained-class}
\alias{BoxConstraints-class}
\alias{LinearConstraints-class}
\alias{getConstraints,Constraints-method}
\alias{setConstraints,Constraints-method}

\title{Parameter constraints}
\description{Objects of this class represent parameter constraints. Currently, two types of constraints are implemented: box constraints, and linear constraints. The class Unconstrained is used when there are no constraints.}

\author{Maarten Speekenbrink}

