\name{mcplR-internal}

\alias{rFl}
\alias{lFr}
\alias{lFr,GCMinterval,GaussianMixtureResponse-method}
\alias{lFr,LearningModel,GlmResponse-method}
\alias{lFr,LearningModel,ResponseModel-method}
\alias{lFr,McplModel,missing-method}
\alias{rFl,McplModel,missing-method}
\alias{rFl,ResponseModel,LearningModel-method}


\alias{rFl-methods}
\alias{lFr-methods}
\alias{mcpl.prepare}
\alias{relist.factor}
\alias{relist.pdMat}
\alias{relist.NULL}
\alias{NTimes-class}
\alias{nTimes}

\alias{fixedListToVec}

\alias{AIC,missing-method}
\alias{BIC,missing-method}

\alias{getReplication-method}

\title{mcplR internal functions, methods and classes.}

\description{
  Internal mcplR functions, methods and classes.
}

\details{
  These are not to be called directly by the user.
}

\author{Maarten Speekenbrink}
