\name{runm-methods}
\docType{methods}
\alias{runm-methods}
\alias{runm,ANY-method}

\title{Run a model over the data.}
\description{
The \code{runm} runs a model over data, i.e., calculates certain aspects of the
model on the basis of the model data.
}
\section{Methods}{
\describe{

\item{object = "ANY"}{Generic function.}

}}
\keyword{methods}