\name{SMPT}
\alias{SMPT}

\docType{data}

\title{Stock Market Prediction Task}

\description{
	This data set contains responses of two participants in the Stock Market 
  Prediction Task (SMPT). In the SMPT, participants make 
	predictions (r) of the change in share price of a fictional company based 
  on the change in share price (c1 and c2) of two related companies. The data 
  are from participants S7 and S10 in Experiment 1 of Speekenbrink \& Shanks 
  (2008). Each participant performed 300 trials. At trials 100 and 200, the 
  relation between the cues (c1 and c2) and criterion (y) underwent sudden 
  changes.
  }

\usage{data(SMPT)}

\format{Data.frame}

\source{

Speekenbrink, M. \&  Shanks, D.R. (2008). Learning in a changing environment. 
\emph{Manuscript submitted for publication}.

}

\keyword{datasets}
