\name{WPT}
\alias{WPT}

\docType{data}

\title{Weather Prediction Task data}

\description{
	This data set contains responses of two participants (one amnesic and one
	control) in the Weather Prediction Task (WPT). In the WPT, participants make
	predictions (r) of the state of the weather (y, 0 = sunny, 1 = rainy), on the
	basis of four ``tarot cards'' (x1-x4, 0 = absent, 1 = present). The WPT
  consists of 200 trials.
}

\usage{data(WPT)}

\format{Data.frame}

\source{

	Speekenbrink, M., Channon, S and Shanks, D.R. (2008). Learning strategies in
  amnesia. \emph{Neuroscience and Biobehavioral Reviews},\bold{32},292-310.

}

\keyword{datasets}
