\name{ConstraintsList-class}
\alias{ConstraintsList}
\alias{ConstraintsList-class}
\alias{BoxContraintsList-class}
\alias{LinConstraintsList-class}
\title{Parameter constraints}
\description{Objects of this class represent parameter constraints. Currently, two types of constraints
are implemented: box constraints, and linear constraints.}
\author{Maarten Speekenbrink}

