\name{mcplr-internal}

\alias{rFl}
\alias{lFr}
\alias{mcpl.prepare}
\alias{relist.factor}
\alias{relist.pdMat}
\alias{relist.NULL}

\title{Mcplr internal functions}

\description{
  Internal mcplR functions, methods and classes.
}

\details{
  These are not to be called by the user.
}

\author{Maarten Speekenbrink}

\keyword{methods}