\name{mcplR-internal}

\alias{rFl-methods}
\alias{lFr-methods}
\alias{mcpl.prepare}
\alias{relist.factor}
\alias{relist.pdMat}
\alias{relist.NULL}
\alias{NTimes-class}
\alias{nTimes}
\alias{ParStruct-class}
\alias{fixedListToVec}

\alias{AIC,missing-method}
\alias{BIC,missing-method}

\alias{getReplication-method}

\alias{getPars,ParStruct-method}
\alias{setPars,ParStruct-method}
\alias{rep,ParStruct-method}

\title{mcplR internal functions, methods and classes.}

\description{
  Internal mcplR functions, methods and classes.
}

\details{
  These are not to be called directly by the user.
}

\author{Maarten Speekenbrink}
